+++ src/modules/Commands.ts
@@ -44,6 +44,10 @@ export default class Commands extends ModuleInstance {
         return true;
     }
 
+    override get CanDisable(): boolean {
+        return true;
+    }
+
     override get SettingsScreen(): ((character: BCPlusCharacter | null) => GUIScreen) | null {
         return (character) => new CommandsScreen(this, character);
     }